requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
mysql-connector-python==9.1.0
